+++ test/JsonApiDotNetCoreTests/IntegrationTests/ResourceDefinitions/Reading/MoonDefinition.cs
@@ -11,7 +11,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.ResourceDefinitions.Reading;
 [UsedImplicitly(ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature)]
 // The constructor parameters will be resolved from the container, which means you can take on any dependency that is also defined in the container.
 public sealed class MoonDefinition(IResourceGraph resourceGraph, IClientSettingsProvider clientSettingsProvider, ResourceDefinitionHitCounter hitCounter)
-    : HitCountingResourceDefinition<Moon, int>(resourceGraph, hitCounter)
+    : HitCountingResourceDefinition<Moon, long>(resourceGraph, hitCounter)
 {
     private readonly IClientSettingsProvider _clientSettingsProvider = clientSettingsProvider;
