+++ test/JsonApiDotNetCoreTests/UnitTests/Serialization/Response/Models/Song.cs
@@ -5,7 +5,7 @@ using JsonApiDotNetCore.Resources.Annotations;
 namespace JsonApiDotNetCoreTests.UnitTests.Serialization.Response.Models;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
-public sealed class Song : Identifiable<int>
+public sealed class Song : Identifiable<long>
 {
     [Attr]
     public string Title { get; set; } = null!;
